fix clippy and create module
